+++ src/app/service/service_worker/subscribe.ts
@@ -15,7 +15,8 @@ import { type Group } from "@Packages/message/server";
 import { type ScriptService } from "./script";
 import type { InstallSource } from "./types";
 import { publishSubscribeInstall, subscribeSubscribeInstall } from "../queue";
-import { checkSilenceUpdate, InfoNotification, ltever } from "@App/pkg/utils/utils";
+import { checkSilenceUpdate, InfoNotification } from "@App/pkg/utils/utils";
+import { ltever } from "@App/pkg/utils/semver";
 import type { ScriptInfo } from "@App/pkg/utils/script";
 import { fetchScriptInfo, prepareSubscribeByCode } from "@App/pkg/utils/script";
 import Cache from "@App/app/cache";
