+++ src/pages/components/ScriptResource/index.tsx
@@ -2,7 +2,7 @@ import type { Resource } from "@App/app/repo/resource";
 import type { Script } from "@App/app/repo/scripts";
 import { ResourceClient } from "@App/app/service/service_worker/client";
 import { message } from "@App/pages/store/global";
-import { base64ToBlob } from "@App/pkg/utils/script";
+import { base64ToBlob } from "@App/pkg/utils/utils";
 import { Button, Drawer, Input, Message, Popconfirm, Space, Table } from "@arco-design/web-react";
 import type { RefInputType } from "@arco-design/web-react/es/Input/interface";
 import type { ColumnProps } from "@arco-design/web-react/es/Table";
